VideoAsset: Add a "wrapper" property on VideoAssets
This property is just the asset ID of another article that "thinly
wraps" the video, in the sense that it provides chrome for the video
and not much else. Applications that want to be able to use content
directly aren't interested in these wrapper articles - they would
prefer to use the underlying media directly.

https://phabricator.endlessm.com/T20761