GalleryVideoArticle: Add 'wraps' property to metadata
This property is just a list of all assets that the GalleryVideoArticle
is a thin wrapper for. The property is a list since it is possible that
a GalleryVideoArticle may wrap more than one thing going forward, even
though it only has a "main video" for now.

Note that this property itself isn't directly used - instead, it is
used by SOMA to tag the article as being a thin wrapper for some
other media object so that we can exclude the wrapper article from
search results when we want to go to the underlying media directly.

https://phabricator.endlessm.com/T20761