Add documentation for libingester public API
This uses documentation.js as a docs generation tool. (The documentation
comment format is compatible with jsdoc, a widely used standard, but the
output is a bit nicer.)

I've copied in all the documentation of API methods that we wrote for
https://github.com/endlessm/libingester/wiki/Converting-1.x-ingester-to-2.x-API
and put stubs for all the other public API.

Currently I'm not treating DictionaryAsset, GalleryVideoAsset, etc., as
public API since they are specific to one or a few apps.

The `npm run docs:build` command will build the docs in a `docs/` folder
locally. You can load `docs/index.html` in your browser to see what is
generated. `npm run docs:publish` will commit the docs folder. It's
possible to set up GitHub Pages to publish from the `docs/` folder on
the master branch, so that's how we'll publish it.

https://phabricator.endlessm.com/T21011