test: Cut down on verbosity of expected exceptions
There are a bunch of tests that expect exceptions to be thrown, and
those exceptions get logged with their whole stack which involves
hundreds of lines of Mocha. If we set the stack to a blank string then
we still get the exceptions logged, but they don't completely destroy
the test output.

https://phabricator.endlessm.com/T20603