handle event 'unhandledRejection', and log it
winston can handle exceptions, but not inside promises. If an
exception happens inside a promise, the promise is rejected. And if
there is no catch() to handle it through the promise chain, Node emits
the 'unhandledRejection' event. Adding a callback to this event, we
can log the error, and also exit the process with a return code 1.