config: Change behaviour of settings passed to Hatch
Merge the options passed to the Hatch constructor with the options
passed in the command line.

The original behaviour was to fully override the command line options
with the options passed to the Hatch constructor.

Added tests for command line options.  The config documentation is now
a bit easier to read.

https://phabricator.endlessm.com/T22232