Make revisionTag optional and allow to override it
Sometimes the revisionTag can be another value rather than just the last
modified date (like the ETag or something like that), let's allow the
ingester to override the revisionTag for the asset.

Also, since the Soma pipeline is adding support to calculate the
revisionTag using hashing, let's allow the ingester to not specify
the revisionTag at all, so it's calculated down the pipeline.